dfekcvbw

+++ variables_if_else_while/100-print_comb3.c
@@ -1,11 +1,10 @@
-// Online C compiler to run C program online
 #include <stdio.h>
 
 int main() {
     int a[10] = {0,1,2,3,4,5,6,7,8,9};
     int b[10] = {0,1,2,3,4,5,6,7,8,9};
     
-    int i,j;
+    long unsigned int i,j;
     for (i = 0; i < sizeof(a) / sizeof(int); i++)
     {
